changes in UserService and UserController
