+++ src/main/java/com/deltabook/services/ContactServiceImpl.java
@@ -45,7 +45,4 @@ public class ContactServiceImpl implements ContactService{
         Contact contact = contactRepository.findByFriendFromIdAndFriendToId(fromUser, toUser);
         contactRepository.delete(contact);
     }
-    public void SaveContact(Contact contact) {
-        contactRepository.save(contact);
-    }
 }
