+++ src/main/java/com/deltabook/services/MessageService.java
@@ -8,6 +8,4 @@ public interface MessageService {
 
     Message sendMessage(User userFrom, SendMessage sendMessage);
 
-public interface MessageService {
-    void SaveMessage(Message message);
 }
