+++ src/main/java/com/deltabook/services/MessageServiceImpl.java
@@ -8,7 +8,6 @@ import com.deltabook.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 @Service
 public class MessageServiceImpl implements MessageService {
