+++ src/main/resources/templates/upload_avatar.html
@@ -9,7 +9,7 @@
 <img src="../static/images/logo.png" width="419" th:src="@{images/logo.png}"/>
 <h1>Загрузка аватарки</h1>
 <a href="/">Назад в панель пользователя</a>
-<form action="/upload" method="post" enctype="multipart/form-data">
+<form action="/upload_avatar" method="post" enctype="multipart/form-data">
     <input type="file" name="files" multiple>
     <p></p><input type="submit" value="Upload Files"></input></p>
 </form>
